requirements.txt
appdirs==1.4.4
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
python-dotenv==0.17.1
python-google-places==1.4.2
regex==2021.4.4
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5
